failing test
running: `cargo test --release  -- tests::test_run_wc --exact --show-output`

yields a fail right now as the ROM used for PublicParams does not match the len of the ROM used for the snark. I think we need to pre-commit to a maximum length of ROM and have a dummy value in ROM that signals to not do anything (e.g., -1).